Struct exempi::XmpFile

pub struct XmpFile { /* fields omitted */ }

Methods

impl XmpFile

fn new() -> XmpFile

Create new XmpFile

fn open_new(p: &str, options: OpenFlags) -> Option<XmpFile>

Create and open a new XmpFile Equivalent to calling new then open. Return None in case of failure

fn open(&mut self, path: &str, options: OpenFlags) -> bool

Open an XmpFile. Usually called after new.

fn close(&mut self, options: CloseFlags) -> bool

Close the XmpFile

fn is_null(&self) -> bool

Return true if native pointer is null

fn get_new_xmp(&self) -> Option<Xmp>

Get a new Xmp fronm the currently open file

fn get_xmp(&self, xmp: &mut Xmp) -> bool

Get the xmp data an Xmp.

fn get_xmp_xmpstring(&self, packet: &mut XmpString, info: &mut PacketInfo) -> bool

Get the xmp packet as a string.

fn can_put_xmp(&self, xmp: &Xmp) -> bool

Return true if it can put the Xmp into the XmpFile.

fn can_put_xmp_xmpstring(&self, xmp_packet: &XmpString) -> bool

Return true if it can put the XmpString packet into the XmpFile.

fn can_put_xmp_str(&self, xmp_packet: &str) -> bool

Return true if it can put the XmpString packet into the XmpFile.

fn put_xmp(&mut self, xmp: &Xmp) -> bool

Put the Xmp into the XmpFile

fn get_file_info(&self, file_path: &mut String, options: &mut OpenFlags, format: &mut FileType, handler_flags: &mut FormatOptionFlags) -> bool

Get info from the XmpFile.

fn check_file_format(path: &str) -> FileType

Check the file format for the specified path

fn get_format_info(format: FileType, options: &mut FormatOptionFlags) -> bool

Get FormatOptions for the FileType