Struct exempi::XmpString

pub struct XmpString { /* fields omitted */ }

The string wrapper from Exempi. It is meant to be used for output parameter. But gives you ownership of the string. Because of the way the C API of Exempi is implemented, we provide this type instead of using std::string::String to avoid copying of strings until needed. They are mostly returned in an Option enum. XmpString support several of the standard traits.

use exempi::{Xmp,PropFlags};

let xmp = Xmp::new();
let mut flags = PropFlags::empty();

if let Some(ref xmpstring) = xmp.get_property("http://rust.figuiere.net/ns/rust/", "rust", &mut flags) {
   println!("property value is {}, flags {}", xmpstring, flags.bits());
   println!("string len: {}", xmpstring.len());
   let s = String::from(xmpstring.to_str());
   println!("converted to std::String: {}", s);
}Run

Methods

impl XmpString

fn new() -> XmpString

Create a new XmpString

fn is_null(&self) -> bool

Native pointer is NULL

fn as_ptr(&self) -> *const XmpString

Return the native pointer

fn as_mut_ptr(&mut self) -> *mut XmpString

Return the mutable native pointer

fn len(&self) -> usize

Return the length of the string

fn to_str(&self) -> &str

Convert to a str